Add support for ClassProxyHints
With this change it is now possible to express a ClassProxyHint
that will cause a class proxy (and supporting auxiliary classes)
to be constructed at application build time and subsequently
loaded and used at runtime. In this way Spring will avoid
attempting to dynamically create and define a class at runtime
which is not allowed in native-images.

These changes are using some of the tech from pull request 1283
against Spring Framework which uses bytebuddy to generate Spring
compatible proxies. The changes have just been massaged so all
generation happens at build time through the Spring Native AOT
system. Then at runtime a substitution is being applied to the
proxy loading mechanism that will load these pre-generated proxies.

Three new test samples are added: classes-proxies-aop,
security-method-class-proxies and batch-io.

The substitution is built such that if a class proxy hint was
missed, it will produce an error message describing the exact
hint that can then be pasted into the application source such
that on next build the proxy will be in place.